wip: split observability support modules
